+++ frontend/site/navigation/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/epochcrysis/web/compare/@site/navigation@0.3.0...@site/navigation@0.4.0) (2019-08-03)
+
+
+### Bug Fixes
+
+* **navigation:** remove useless imports ([86e7349](https://github.com/epochcrysis/web/commit/86e7349))
+
+
+### Features
+
+* **frontend:** add mobile navigation ([f8c6abd](https://github.com/epochcrysis/web/commit/f8c6abd))
+
+
+
+
+
+# 0.3.0 (2019-08-03)
+
+
+### Features
+
+* **frontend:** add navigation ([3eb3e21](https://github.com/epochcrysis/web/commit/3eb3e21))
+
+
+
+# 0.2.0 (2019-07-26)
+
+
+### Features
+
+* **common:** init ([3071609](https://github.com/epochcrysis/web/commit/3071609))
+
+
+
+
+
 # 0.2.0 (2019-07-26)